requirements.txt
garminconnect
streamlit
pandas
plotly
statsmodels
openai
fpdf2
supabase
